// utils/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      MenuItems: {
        Row: {
          created_at: string
          id: number
          is_addon: boolean
          name: string | null
          per_serving: number | null
          price: number | null
          type: Database["public"]["Enums"]["Food_type"] | null
        }
        Insert: {
          created_at?: string
          id?: number
          is_addon?: boolean
          name?: string | null
          per_serving?: number | null
          price?: number | null
          type?: Database["public"]["Enums"]["Food_type"] | null
        }
        Update: {
          created_at?: string
          id?: number
          is_addon?: boolean
          name?: string | null
          per_serving?: number | null
          price?: number | null
          type?: Database["public"]["Enums"]["Food_type"] | null
        }
        Relationships: []
      }
      OrderItems: {
        Row: {
          created_at: string
          id: number
          menu_item_id: number | null
          name: string | null
          order_id: number | null
          price: number | null
          protein: number | null
          proteinName: string | null
          proteinPerServing: number | null
          proteinPrice: number | null
          proteinQty: number | null
          swallow: number | null
          swallow_price: number | null
          swallowName: string | null
          swallowQty: number | null
          total: number | null
        }
        Insert: {
          created_at?: string
          id?: number
          menu_item_id?: number | null
          name?: string | null
          order_id?: number | null
          price?: number | null
          protein?: number | null
          proteinName?: string | null
          proteinPerServing?: number | null
          proteinPrice?: number | null
          proteinQty?: number | null
          swallow?: number | null
          swallow_price?: number | null
          swallowName?: string | null
          swallowQty?: number | null
          total?: number | null
        }
        Update: {
          created_at?: string
          id?: number
          menu_item_id?: number | null
          name?: string | null
          order_id?: number | null
          price?: number | null
          protein?: number | null
          proteinName?: string | null
          proteinPerServing?: number | null
          proteinPrice?: number | null
          proteinQty?: number | null
          swallow?: number | null
          swallow_price?: number | null
          swallowName?: string | null
          swallowQty?: number | null
          total?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "OrderItems_menu_item_id_fkey"
            columns: ["menu_item_id"]
            isOneToOne: false
            referencedRelation: "MenuItems"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "OrderItems_order_id_fkey"
            columns: ["order_id"]
            isOneToOne: false
            referencedRelation: "Orders"
            referencedColumns: ["id"]
          },
        ]
      }
      Orders: {
        Row: {
          created_at: string
          id: number
          order_summary: string | null
          payment_method: string | null
          payment_status: boolean | null
          total_amount: number | null
        }
        Insert: {
          created_at?: string
          id?: number
          order_summary?: string | null
          payment_method?: string | null
          payment_status?: boolean | null
          total_amount?: number | null
        }
        Update: {
          created_at?: string
          id?: number
          order_summary?: string | null
          payment_method?: string | null
          payment_status?: boolean | null
          total_amount?: number | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      Food_type: "soup" | "swallow" | "protein"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never